+++ plugins/linear-genome-view/src/LinearGenomeView/components/LinearGenomeView.tsx
@@ -11,7 +11,7 @@ import TrackContainer from './TrackContainer'
 import TracksContainer from './TracksContainer'
 import ImportForm from './ImportForm'
 import MiniControls from './MiniControls'
-import SequenceDialog from './SequenceDialog'
+import GetSequenceDialog from './GetSequenceDialog'
 import SearchResultsDialog from './SearchResultsDialog'
 
 type LGV = LinearGenomeViewModel
@@ -62,9 +62,9 @@ const LinearGenomeView = observer(({ model }: { model: LGV }) => {
   return (
     <div style={{ position: 'relative' }}>
       {model.seqDialogDisplayed ? (
-        <SequenceDialog
+        <GetSequenceDialog
           model={model}
-          handleClose={() => model.setSequenceDialogOpen(false)}
+          handleClose={() => model.setGetSequenceDialogOpen(false)}
         />
       ) : null}
       {model.isSearchDialogDisplayed ? (
