+++ plugins/linear-genome-view/src/LinearGenomeView/components/SequenceSearchDialog.tsx
